backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
celery
redis
requests
python-dotenv
alembic
openai
langchain